+++ Changelog.md
@@ -3,6 +3,12 @@
 	<li>Migrate golang.org/x/crypto to standard library imports (https://github.com/golang/go/issues/65269)</li>
 </ul>
 
+# v1.46 (Released 01/29/2025)
+<ul>
+	<li>✓ Added Picocrypt version to the window title</li>
+	<li>✓ Added ability to automatically unzip archives upon decryption</li>
+</ul>
+
 # v1.45 (Released 12/05/2024)
 <ul>
 	<li>✓ Bumped GitHub Actions Ubuntu 22 -> 24 and macOS 14 -> 15</li>